+++ packages/jet_print/lib/src/designer/l10n/jet_print_localizations_de.dart
@@ -319,6 +319,12 @@ class JetPrintLocalizationsDe extends JetPrintLocalizations {
   String get reportLabel => 'Bericht';
 
   @override
+  String get propertiesName => 'Name';
+
+  @override
+  String get reportNameHint => 'Berichtsname';
+
+  @override
   String get propertiesEmptyHint =>
       'Wählen Sie ein Objekt, um seine Eigenschaften zu bearbeiten.';